added pictures
in the simulation picture there is the simulation for 3 colours (red, green and blue). I changed the representation of the colors from decimal to binary representation and the row and column representation from decimal to hexadecimal representation.

The other image shows the files contained in my project. I moved  the test bench from the Design source to the simulation sources.